+++ .gitignore
@@ -69,3 +69,6 @@ nb-configuration.xml
 # Ignore built resources to display in package
 src/main/resources/META-INF/resources/*
 secret.properties
+
+#environment variables
+.env
