+++ src/main/js/src/components/Pages/CompanyInformation/CompanyInformationCompany.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import CustomSelectWrapper from '../Inputs/CustomSelect/CustomSelectWrapper';
-import DefaultSelect from '../Inputs/CustomSelect/DefaultSelect';
-import CustomAsyncSelect from '../Inputs/CustomSelect/CustomAsyncSelect';
-import Input from '../Inputs/Input';
-import { formField } from '../formModels/formFieldModel';
+import CustomSelectWrapper from '../../UIComponents/Inputs/CustomSelect/CustomSelectWrapper';
+import DefaultSelect from '../../UIComponents/Inputs/CustomSelect/DefaultSelect';
+import CustomAsyncSelect from '../../UIComponents/Inputs/CustomSelect/CustomAsyncSelect';
+import Input from '../../UIComponents/Inputs/Input';
+import { formField } from '../../UIComponents/FormComponents/formModels/formFieldModel';
 import { companies } from '../../../Constants/Constants';
 
 /**
