+++ src/main/js/src/components/Pages/MembershipLevel/MembershipLevelFeeTable.js
@@ -4,7 +4,7 @@ import React from 'react';
  * This is a pure static html table copied from membership site:
  * https://www.eclipse.org/membership/#tab-fees
  */
-const MembershipFeeTable = () => {
+const MembershipLevelFeeTable = () => {
   return (
     <>
       <div className="vertical-align margin-top-50">
@@ -96,4 +96,4 @@ const MembershipFeeTable = () => {
   );
 };
 
-export default MembershipFeeTable;
+export default MembershipLevelFeeTable;
