+++ src/main/js/src/components/UIComponents/FormPreprocess/FormWrapper.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import MultiStepForm from '../FormComponents/MultiStepForm';
-import MembershipContext from '../../Context/MembershipContext';
-import SignIn from '../SignIn/SignIn';
+import MembershipContext from '../../../Context/MembershipContext';
+import SignIn from '../../Pages/SignIn/SignIn';
 
 const FormWrapper = () => {
   const { currentUser, currentFormId } = useContext(MembershipContext);
