+++ src/main/js/src/components/UIComponents/Steppers/stepHelpers.js
